requirements.txt
chromadb==1.5.5
google-genai==1.68.0
numpy==2.4.3
python-dotenv==1.2.2
